MB-63024 Handle generated nested aggregates ...    ... if FROM clause subquery is planned multiple times
The problematic query has a FROM clause subquery that is used
under the inner of a nested-loop join, in which case we put
an ExpressionScan on top of the Subquery. In the subquery
plan we can do aggregate pushdown, which for a COUNT aggregate
we convert into SUM(cover(COUNT)) if the pushdown is partial,
i.e. two level aggregation is required.

The issue here is that with join enumeration, we end up
planning the FROM clause subquery multiple times, with the
above-mentioned rewriting of nested aggregates. When the
subquery gets planned and executed at execution time, the
planner sees the nested aggregates and returns an error.

Ideally we should be able to plan the FROM clause subquery
only once and save the plan for future use, however
previous fix for this only handles multiple planning of
subqueries at execution time. There is no simple way to
save the plan generated at planner phase and use it at
execution time (the subquery plans are saved in execution
context). Instead, we make multiple planning possible in
this case.

The fix includes the following:

 - add a new Aggregate flag to mark the SUM aggregate as
   "generated"
 - when rewriting COUNT to SUM(cover(COUNT)), mark the new
   SUM aggregate as "generated"
 - during checking for nested aggregate, skip the generated
   aggregate (avoid generating error)
 - as part of the covering transformation with covered index
   aggregate, make a copy of the original aggregate expression
   before changing it, such that if the same expression is
   rewritten again, we don't accidentally generate a nested
   aggregate like COUNT(cover(COUNT))

Change-Id: I969d95631d9cdbf296c0ff5624e8bf7a95fa5e29
Reviewed-on: https://review.couchbase.org/c/query/+/213968
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>