MB-63057 Consider index key that's also used in ...    ... index condition when comparing indexes with RBO
When the rule-based optimizer (RBO) compares two indexes, it uses
a variety of criteria for comparison purpose. For this bug, the
two indexes are both "favorable" in one of the areas being
compared. One index has a WHERE clause (the other does not), but
the other index has less number of total keys. Both indexes are
being kept and an intersect scan is being done.

We now add an extra area to be considered. We already have nEqCond,
which represents the number of equality predicates in the index
condition. However current logic only adds to nEqCond if the same
field is not already in index key. Thus we now also count the
number of index keys that has a corresponding equality predicate
in index condition, i.e. what was previously ignored now gets a
separate counter. This new counter is then used for comparing
indexes (before comparing number of index keys) such that with
others being equal, we favor an index whose new counter is larger,
i.e. the index has an index condition with an equality predicate
that was previously unaccounted for.

The logic for getting number of equality predicates in index
condition is also modified. Previous code counts any equality
predicate, even if it is embedded deeply inside some other
expressions. Now we only consider "top-level" equality predicates,
i.e. either a single EQ predicate, or any top-level EQ predicate
under an AND. A third addition is to mark that the index condition
has no other extra predicates.

Change-Id: I8cfb6ea2c7263712d3b962135b4baf4028d3f2ab
Reviewed-on: https://review.couchbase.org/c/query/+/214258
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>