MB-63147 Proper handling of correlated CTE reference in inline UDF
A UDF is defined with two CTEs, the second one is correlated since it
references function variables. When the function is executed the parser
gets a "CORRELATED <subquery>" expression and we have logic in place
to formalize the subquery expression to properly mark for correlation.
The issue here is that the correlation reference is in the definition
of a CTE which current code does not properly handle when we are trying
to add the correlation references from the subquery to the parent.

We already have logic to handle function variables, the fix is to add
the same handling for CTE references as well.

Change-Id: Ic450b4b24658219182249389a16134aca76edf14
Reviewed-on: https://review.couchbase.org/c/query/+/213795
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker