MB-63163: [BP 7.2.6]  Make sure arrayId is properly assigned ...    ... for ON-clauses of outer joins
This bug is a regression from the fix for MB-61549.

The fix for MB-61549 introduced the concept of "arrayId", where
an array predicate (ANY, EVERY, ANY AND EVERY) is not marked with
an array id, such that at index span generation time we can
properly identify index spans generated from the same array
predicate.

The issue with this bug is that if an array predicate is in the
ON-clause of an outer join, then the previous logic for assigning
array id missed the ON-clause and thus we trigger a sanity error
with a '0' value array id.

The intent of the original fix is to assign an array id early on
in the planning process. The fix contains the following:

 - add logic to assign array id for ON-clause filters when we
   traverse the FROM clause looking for keyspace references. This
   applies to ANSI JOIN and ANSI NEST. Assign the array id for both
   outer and inner join.
 - modify existing logic in processPredicateBase(), to only assign
   array id if the predicate passed in is NOT a ON-clause predicate
 - add logic in MERGE processing to also assign array id for
   ON-clause predicates (outer or inner)
 - the processing for DELETE/UPDATE does not change (previous fix)

Change-Id: I2a156216e16023e8a20ca78ed8a4ed23e5f5b09e
Reviewed-on: https://review.couchbase.org/c/query/+/214566
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/214614