MB-63204 Provide optional (opt-in) mode to ...
automatically re-prepare prepared statements when indexer metadata
versions change.

Also links sequential scans to the GSI indexer if there is one.
The idea here is that the SequentialScan indexer will report the GSI
indexer's metadata version as its own, so that any GSI index changes
trigger changes for anything using the sequential scan indexer.  The
second part of this is to report the sequential scan index ID as a
string including the metadata version.

This means that when an index is created - and the mode is enabled - a
prepared statement using the sequential scan indexer gets re-prepared.
(It is still free to select the sequential scan of course.)  In the
absence of GSI changes, plans using sequential scans will not be
re-prepared.

Another part of this is to indicate a re-prepare is desired when a
prepared statement's indexer metadata version changes.  The idea for
this part is that we can automatically switch between indexes within an
indexer should a more suitable one be created.

To opt in, the relevant feature control bit must be set in the cluster
configuration.

Also provide a means to manually trigger a re-prepare on next execution.
This is achieved by permitting an UPDATE on system:prepareds - the only
action considered in the update is UNSETting planPreparedTime; all other
actions are reported as errors (unsetting a field that doesn't exist is
a no-op so is silently ignored).

Change-Id: If48a85a2ded9ef1254a37d3037a9377faac4b128
Reviewed-on: https://review.couchbase.org/c/query/+/214692
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>