MB-63223 Add xattrs to INSERT & UPSERT options...
... and support xattr modifications via UPDATE.

`xattrs` is effectively a containing document for which we only ever
have access to the content.  (We can't replace the document itself.)

For INSERT and UPSERT we can only set top-level xattr values to those
specified in the statement options.  We can't delete fields nor modify
only nested elements.

For UPDATE, we fetch the existing xattr values referenced in the
mutation clauses so that we can manipulate nested fields in isolation.

Change-Id: I7e2b4dfba81b324d79a989715580313291389b19
Reviewed-on: https://review.couchbase.org/c/query/+/215072
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>