MB-63267 Restore group before considering primary scan
If no secondary index scan is available for a query, the planner
considers primary scan when appripriate. Before doing so, it restores
this.order and conditionally this.limit/this.offset such that the
primary scan can perform certain index pushdown when appropriate.
The issue with the problematic query is that it also has GROUP BY
specified, however since the code did not restore this.group, the
information is "lost" in a way and we incorrectly determined that
LIMIT can be pushed down.

The fix is to also restore this.group such that the primary index
scan can make proper decision regarding index pushdown.

Change-Id: Ie488f9f1d13d81343d088870607eaa0c36c3b833
Reviewed-on: https://review.couchbase.org/c/query/+/214880
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/214924