MB-63274 Do not add Distinct operator in FROM clause subquery
For a Set Operation that requires distinct (UNION/INTERSECT/EXCEPT),
we add a Distinct operator in each arm of the set operation. This is
achieved by setting this.setOpDistinct indicator in the builder.
This indicator is set to true right before we plan each arm of the
set operation. The problem is that if there is a FROM clause subquery,
it also inherits this indicator and thus adds a Distinct operator for
the FROM clause subquery, which is incorrect.

The fix is to check whether a Select is under a set operation, and
if not, reset this.setOpDistinct to false. This requires we save and
restore this.setOpDistinct around the planning of a Select.

Change-Id: I635a28ea0393766ee17c7ff2f041e99b02d9ca17
Reviewed-on: https://review.couchbase.org/c/query/+/214895
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/214925