MB-60680 Allow ANN expression as projection alias and LET variable
Currently an ANN expression is only properly recognized if it is specified
in the ORDER BY clause. We now add ability to also use ANN expression in
projection clause with a projection alias, and use the projection alias in
the ORDER BY clause. Similarly, we also allow ANN expression in a LET
clause, and use the LET variable in the ORDER BY clause.

Note that in both cases, we only support the simple case where an identifier
is used in the ORDER BY clause, and the identifier is either a projection
alias or a LET variable. We don't attempt to further check whether the
identifier is embedded in the ORDER BY expression. Since when the ANN
expression is used in the ORDER BY clause directly, we only support the case
where the ANN expression itself is directly used in the ORDER BY clause, we
don't check whether the ANN expression is embedded somewhere inside the
ORDER BY expression.

Change-Id: If2043e3d77acf06cf155e4df9813afec6781f081
Reviewed-on: https://review.couchbase.org/c/query/+/214977
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>