MB-63311, MB-60680 Add early bailout for index partition set for BHIVE indexes
For BHIVE indexes the planner tries to obtain index partition set to
enable partition elimination for BHIVE indexes. The issue with this
bug is that we currently do not have logic to handle early bailout,
i.e. if the index is not partitioned, or if there is no predicates.

The fix is to add logic for early bailout.

Change-Id: I465d150c440e60bbbd9371ac1bc02bd1aadd3f00
Reviewed-on: https://review.couchbase.org/c/query/+/215028
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>