MB-63359 Avoid multiple CREATE PRIMARY INDEX calls during migration
At the end of the migration process, after all CBO stats and
scope functions have been moved to the new _query collection
under each bucket, we attempt to create a primary index on
the _query collection. This should in theory happen once for
each bucket for each migration process (UDF and CBO stats),
however due to presence of MB-63078 it is possible to have
this called repeatedly which results in many "duplicated
index" error in the query log.

While we cannot completely eliminate the duplicated calls
(e.g. if there are multiple query nodes we cannot easily
coordinate between query nodes) we can at least make sure
that on each query node, for each migration process, we
only attempt to call CREATE PRIMARY INDEX for each bucket
once. This is achieved by adding an extra marker on the
migration bucket structure, and only call the create if
the marker is not set yet.

Change-Id: I2e436b36c9d07d7e914e6c1dfbc2aa36182433c4
Reviewed-on: https://review.couchbase.org/c/query/+/215204
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>