MB-63223 Revise implementation to...
... ignore virtual XATTRs in OPTIONS.  This also ignores those that have
been fetched for an UPDATE statement so that we don't try to mutate
them (which would result in an error).

This specifically corrects the case:

    SET meta().expiration = meta(o).expiration + nnn

where expiration is fetched as the xattrs.`$document`.exptime virtual
XATTR and means that OPTIONS passed in INSERT or UPSERT that include
virtual XATTRs no longer raise errors because of them.

Also corrected the error processing around processing sub-doc
multi-mutation operation errors to avoid a possible panic.

Change-Id: Id8f5cf489f0399cbeee184269473079169fad57a
Reviewed-on: https://review.couchbase.org/c/query/+/215325
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>