MB-63418 MB-60680 Always copy vector filters when copying baseKeyspaces
Vector filters represent the ANN predicate in the ORDER BY clause of the
original query. When considering OR predicates we make a copy of the
baseKeyspace. The issue is we did not properly copy vector filters when
copying baseKeyspace resulting in not properly detecting that a vector
filter exists.

The fix is to always copy vector filters regardless of whether other
filters are being copied.

Change-Id: Ice59348f3691f3093f73a3636d75edb02f97b1eb
Reviewed-on: https://review.couchbase.org/c/query/+/215661
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>