MB-63420 Track references used in aggregates
When an aggregate includes a reference to a source that may be pooled
type, e.g.  MAX([d.date, d][1]), we need to ensure that "d"'s use in the
aggregate is tracked so when the source item is released "d" remains.
We need also to release references to prior values of "d" when updating
the aggregate to maintain the efficiency of the pooling.

We also have to ensure tracking percolates through container Value types
else the more common constructs such as MAX([d.date, d])[1] don't
maintain the reference counting correctly.

Change-Id: I8749a06046c4fea17e4fbb4f2daf83c27c005763
Reviewed-on: https://review.couchbase.org/c/query/+/215678
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>