MB-63462 MB-60680 Do not use CBO for vector index
The issue with this bug is that we have two qualifying vector indexes
to satisfy the query, both covered. We have proper optimizer statistics
and thus CBO kicks in. It so happens that the "bad" index in this case
has a better resident ratio than the "good" index, and since currently
we do not properly cost the scan of a vector index key, we simply use
the regular cost of index scan, which is actually wrong in this case
since the "bad" index ends up scanning a much larger section of the
index. We use the equivalent of WHOLE_SPAN to represent the vector
index key but in reality we only scan a subset of the centroids.

The fix is to turn off CBO for vector indexes for the initial
release. We'll need to wait till we have proper costing for vector
index scans to properly use CBO when comparing 2 vector indexes.

Change-Id: I5a4cd4d9ccd66f2612ed1127e0933ccd3db3797e
Reviewed-on: https://review.couchbase.org/c/query/+/215931
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>