MB-63459 [BP 63223] Add xattrs to INSERT & UPSERT options...
... and support xattr modifications via UPDATE.

`xattrs` is effectively a containing document for which we only ever
have access to the content.  (We can't replace the document itself.)

For INSERT and UPSERT we can only set top-level xattr values to those
specified in the statement options.  We can't delete fields nor modify
only nested elements.

For UPDATE, we fetch the existing xattr values referenced in the
mutation clauses so that we can manipulate nested fields in isolation.

Includes implementation revision to...

... ignore virtual XATTRs in OPTIONS.  This also ignores those that have
been fetched for an UPDATE statement so that we don't try to mutate
them (which would result in an error).

This specifically corrects the case:

    SET meta().expiration = meta(o).expiration + nnn

where expiration is fetched as the xattrs.`$document`.exptime virtual
XATTR and means that OPTIONS passed in INSERT or UPSERT that include
virtual XATTRs no longer raise errors because of them.

Also corrected the error processing around processing sub-doc
multi-mutation operation errors to avoid a possible panic.

Includes adding privs for system XATTR mutations.

Change-Id: I973112778bf78f9de2ba68854201b3903e3d907a
Reviewed-on: https://review.couchbase.org/c/query/+/216004
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>