MB-63479 Provide a mechanism to redirect existing FunctionName ...    ... metaEntry to the new systemEntry after migration
The migration process will move metadata for scope UDFs from its current
location in metakv to the new location in system collection. The current
design for a UDF name is a structure that points to either metakv or
system collection. The issue with this bug is that if a prepared statement
references any scope UDF, then in the prepared plan it has function name
reference which points to metakv. After migration, this name reference is
no longer valid, and an attempt to load the function into function cache
results in a "missing entry" (since it is not found in metakv) and thus
we conclude that the function does not exist.

A couple of issues exist with regard to this process, first, a UDF name
reference should not be allowed to proceed in any action (load, save, etc)
while migration in ongoing. The current mechanism for blocking actions
only work if we need to determine whether metaStorage or systemStorage
should be used. In case of a pre-existing metakv entry, no such test is
performed and thus no blocking is happening; second, after migration
completes the pre-existing metakv entry is no longer valid since all
relevant metadata has been moved. Note that a prepared statement is just
one instance of a pre-existing metakv entry reference. We could also have
a request that generates a UDF name reference before migration kicks in
(i.e. a metakv entry is generated for function name), but the request
persists during/after migration, in which case the same problem can
occur as well.

The fix has the following aspects:

 - during server initialization set up a function pointer that points to
   the function storage.UseSystemStorage(), which is the function used
   to check migration status and block when necessary. This setup is
   required to work around circular dependencies among packages
 - for several actions with a FunctionName, if it is a scope UDF, add
   logic to first call the function pointer setup above, such that it
   can block during migration process. This is needed for the following
   operations involving a function: save/load/execute.
 - during function name construction, when constructing a metakv entry,
   also generate a "dummy" system entry, which serves as a redirect
   link later on. In addition, also adds a boolean indicator in metakv
   entry to indicate whether a redirection is needed
 - a new SetUseSystem() function is added to FunctionName interface, and
   it is only implemented for metakv entry (no-op for all other types).
   This function is called when the above-mentioned wait for migration
   returns (at which point we know migration has completed), to indicate
   that from that point on, the metakv entry needs to be redirected to
   a system entry, and thus use the corresponding system entry
   functionality for load/store/delete/change_counter_management.

Change-Id: I06b8b524df53b500cf5eb5da52ebe6c3427e33e9
Reviewed-on: https://review.couchbase.org/c/query/+/216153
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>