MB-63518 Ensure completed requests streaming...
... stop can be issued for the feature when it is already stopped.

Specifically we weren't exiting when detecting the feature wasn't active
and so proceeded to attempt to signal a non-existent worker routine to
stop.  This was done under lock without a timeout leading to a hang.
(Unfortunately the write to a nil channel didn't panic rather hung else
we may have picked this up sooner.)

The correction is to not only exit if it isn't active without further
actions, but to add a timeout to the channel write allowing the stop to
abort (and release the lock) if there is an unforeseen issue.

Further to this, the statistics presented in vitals information are no
longer gathered under the vitals lock, to limit possible impact of
issues like this (i.e. a bit of a safety-net).

Change-Id: I57b1b7771bf4fc117f0dfeddd2c25fb395c96cc3
Reviewed-on: https://review.couchbase.org/c/query/+/216218
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>