MB-63532 Copy document body out of scan buffer...
... when using "with docs" scans.  The buffer may be reused corrupting
earlier slices that still point into it.

Also change the inferencer to iterate over the result from
value.Fields() rather than using the Actual() interface as this allows
us to easily handle the case when a parsed value has yet to be unwrapped
and unwrapping results in a binaryValue rather than objectValue type.
(Removes the chance of panicking here for a similar reason.)

Change-Id: Ic9f0a3cf89d0a8925bd54121c482fac1fa17d9db
Reviewed-on: https://review.couchbase.org/c/query/+/216299
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>