MB-63544 Undo part of Order optimization
As part of the change for MB-61760, an optimization is introduced
for Order when limit/offset is included. A max heap is used in
this case, and if there is no spilling, then we know the items
are already in reversed order due to max heap processing. In such
cases we simply read the heap array backwards to get the items
in order.

The issue with this optimization is that it does not take into
consideration the sort order of items if several items sort the
same (same sort keys). E.g., if 3 items sort the same, as
  item1, item2, item3
then prior to the optimization we would have resorted the items to
  item3, item2, item1
in which case if an offset is used, we would apply the offset on
the new order and get the results after the offset. With the new
optimization though, we would apply the offset while reading in
reverse order, which will result in the same item (item1) being
returned with different offset value. While this is technically
not completely wrong (since we do not guarantee order if the
sort keys are exactly the same), it is a different behavior than
previous releases (e.g. 7.6.1).

For now we disable this optimization such that the behavior is
consistent with previous releases. It should be possible to still
keep this optimization while maintaining previous behavior, that
is left as future enhancement.

Change-Id: Ibfa6ee68d25b8a1bcf1ffc88263c05c7ebedf1bc
Reviewed-on: https://review.couchbase.org/c/query/+/216325
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker