MB-63496 MB-60680 Retain vector index span when number of ...    ... index spans generated exceeds maximum
The planner has a limit on how many index spans to generate for an
index scan (8192). When there are multiple IN list predicates exist
on several index keys, it is possible to exceed this limit, at which
point the planner simply stops and discard the spans for the rest of
the index keys. For this defect we have a vector index defined with
the vector key being the 7th index key, and there are IN-list
predicates on index keys 2-6, which when fully expanded will exceed
the limit. Current logic then removes the index span for the vector
index key, which causes issues later on since the code expect an
index span for the vector index key.

The fix is to detect early on that a vector index key exists, and
in such situations, instead of discarding the remaining index spans,
we'll keep the index spans for at least as far as the vector index
key. In this case instead of expanding the IN-list predicates,
we'll try to generate a single index span with min/max value from
all values in the IN-list, this then will not cause the number of
index spans to increase, but will still keep (meaning tagging
along with the remaining index keys) the index spans for the
remaining index keys. In situations where we do not know the values
of all the elements of the IN-list in question (e.g. host variables)
we'll simply generate a _WHOLE_SPAN for the index key.

Note that we also have a mechanism to dynamically expand the IN-list
elements if the entire IN-list is specified as a host variable.
However that logic does not need to be changed since in that case
the entire IN-list is represented as a single span with
array_min/array_max as the LOW/HIGH key. The dynamic expansion of
the spans already observes the maximum number of spans (8192) and
will stop expansion once that limit is reached.

Change-Id: I95e300d69d8195547ec07ef8aa94e9b46dd12994
Reviewed-on: https://review.couchbase.org/c/query/+/216367
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>