MB-63578 Restore index pushdown for primary index consideration
This bug is a regression from the fix for MB-63267, although
the underlying issue exists previously.

If optimizer hints is specified for a query the planner can
potentially go through index selection twice, first with
the indexes specified in the hints, and if nothing is returned
then it considers the remaining indexes. As part of the index
selection process we consider primary scan if a secondary
index is not available. Before primary scan consideration we
restore certain index pushdown properties in the planner such
that the primary scan can consider index pushdown as well.

The issue with this process is that when we restore index
pushdown properties for consideration of primary scan, we
currently only restore a subset of the index pushdown
properties. This could result in inconsistencies since in
case a primary scan is not available, we may go through
index selection again (with remaining indexes) with an
inconsistent set of index pushdown properties. Specifically
for this bug, we are missing this.aggs during the second
round of index selection.

The fix contains the following:

 - for save/restore of index pushdown properties, use the
   existing functions that handles all index pushdown
   properties together
 - if considering optimizer hints, check whether any
   primary index is specified in the hints, if not, no need
   to go through primary scan consideration
 - after consideration of primary scan, if no primary scan
   is generated, restore index pushdown properties again
   such that next round of index selection has the original
   index pushdown properties

Change-Id: Ibfb12ef8362d61566ea2e8b045d8651803ea3ad9
Reviewed-on: https://review.couchbase.org/c/query/+/216448
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/216492
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>