MB-63593 Check filter references to keyspaces before pushing filter ...    ... to unnest scan
We added the capability to push filters on an unnest alias to the
unnest scan in MB-56616. The code kind of assumes that there is a
single unnest of the original primary keyspace. In this bug we have
two unnests on the primary keyspace, and we incorrectly pushed a
filter on the second unnest alias to an unnest scan corresponding
to the first unnest alias.

The fix is to add a check for keyspace references in an unnest
filter, such that if any reference does not match the unnest
aliases (could be multiple in case of nested-unnests), we do not
try to push that filter to the unnest scan.

Change-Id: I7d029a88d6e3b8dcfb55ea5de564e3ddfcc99849
Reviewed-on: https://review.couchbase.org/c/query/+/216526
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker