MB-63647 Handle all aggregates in FROM clause subquery ...    ... if subquery is planned multiple times
This bug is very similar to MB-63024 except the aggregate used
in the subquery is MAX while the aggregate used in MB-63024 is
COUNT.

The previous fix for MB-63024 is specific for COUNT since I
mistakenly assume the issue only exists for COUNT. When a
GROUP BY can be partially pushed down to an index scan, we
do two-level aggregation where the indexer performs the
partial aggregation and query does a second level aggregation
on the result of index scan. As part of this process, a
COUNT aggregate is transformed to something like
SUM(cover(COUNT)) where the innter part (cover(COUNT)) is
performed at the indexer and the outer part (SUM) is
performed at query service.

It turns out the same process also applies to other forms of
aggregation as well. E.g., a MAX aggregate is transformed to
MAX(cover(MAX)) in the same fashion, with the cover part
performed at the indexer and the outer part performed at the
query service. Thus the same issue exists for other aggregates
as well when a subquery is planned multiple times (e.g. with
join enumeration).

The fix is to extend the fix for MB-63024 such that it covers
all aggregates instead of just COUNT. The code specific for
COUNT/SUM is also removed.

Change-Id: I7ff7df85a298dff8f8f9e0e5b794ddc986bb3164
Reviewed-on: https://review.couchbase.org/c/query/+/216687
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>