MB-63673 Ensure that one side of a join predicate only has ...    ... a single keyspace reference
When hash join is considered we gather equality join predicates and
try to separate into the build side and the probe side, in which
process we check for keyspace reference to the right-hand side
keyspace. The implicit assumption made here that we will have a join
predicate in the form of "<left-expression> = <right-expression>"
where each side of the equality predicate references one of the
keyspaces being joined. The issue with this bug is that the join
predicate is in the form of "<left-right-expression> = <constant>"
i.e. on one side of the equality both keyspaces are referenced.
This is not properly detected and we end up putting the left
hand side of this filter as a build/probe expression which is
incorrect.

The fix is to tighten the check such that we check for a single
keyspace reference for the left or right side of the equality
predicate, and only allow hash join to proceed if we can still
find a valid equality join predicate.

Similar change is also needed when we attempt to derive a
IS NOT NULL predicate based on a join predicate.

Change-Id: I807da73b198b7acbccb6b31bd7af215bc84c27c7
Reviewed-on: https://review.couchbase.org/c/query/+/216759
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker