MB-63673 Modification of previous fix
The previous fix changed the way we derive IS NOT NULL predicate
based on a join filter. If one side of the join filter references
more than a single keyspace, we no longer derive IS NOT NULL
predicate based on that. This may result in situations where we
could use a secondary index before but not anymore.

Modify the previous fix such that in case we have one side of a
join filter referencing more than a single keyspace, we still
derive a IS NOT NULL filter, but instead of using the original
term (which is wrong since it references more than a single
keyspace), we simply use the leading index key.

Change-Id: Ic8b7859c56447ab6b3de6949b2b3de6b0d24615e
Reviewed-on: https://review.couchbase.org/c/query/+/216772
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>