MB-63673 Ensure that one side of a join predicate only has ...    ... a single keyspace reference
When hash join is considered we gather equality join predicates and
try to separate into the build side and the probe side, in which
process we check for keyspace reference to the right-hand side
keyspace. The implicit assumption made here that we will have a join
predicate in the form of "<left-expression> = <right-expression>"
where each side of the equality predicate references one of the
keyspaces being joined. The issue with this bug is that the join
predicate is in the form of "<left-right-expression> = <constant>"
i.e. on one side of the equality both keyspaces are referenced.
This is not properly detected and we end up putting the left
hand side of this filter as a build/probe expression which is
incorrect.

The fix is to tighten the check such that we check for a single
keyspace reference for the left or right side of the equality
predicate, and only allow hash join to proceed if we can still
find a valid equality join predicate.

Similar change is also needed when we attempt to derive a
IS NOT NULL predicate based on a join predicate.

MB-63673 Modification of previous fix

The previous fix changed the way we derive IS NOT NULL predicate
based on a join filter. If one side of the join filter references
more than a single keyspace, we no longer derive IS NOT NULL
predicate based on that. This may result in situations where we
could use a secondary index before but not anymore.

Modify the previous fix such that in case we have one side of a
join filter referencing more than a single keyspace, we still
derive a IS NOT NULL filter, but instead of using the original
term (which is wrong since it references more than a single
keyspace), we simply use the leading index key.

Change-Id: Ic8b7859c56447ab6b3de6949b2b3de6b0d24615e
Reviewed-on: https://review.couchbase.org/c/query/+/216772
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/216778