MB-63656 MB-60680 Fix issues with covering
If a vector index has include columns defined, then currently
we are not handling the include columns properly in covering
index, which then results in covering transformation not
properly done for include columns.

Also for index key names passed to indexer we are reporting
them as _index_key() even for covering index scans.

The fix:

1. make sure include columns are included in covering info
2. distinguish index key names for covering and non-covering
   indexes
3. simplify a condition introduced in previous checkin

Change-Id: I4248ff4ab135bfdcded5f84a6792c211856f60c9
Reviewed-on: https://review.couchbase.org/c/query/+/216869
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>