MB-63656 MB-60680 Fix issues with order pushdown
If a vector index has include columns defined, then currently
we are not handling the include columns properly in index
pushdown determination. Specifically, we are not considering
include columns when determining whether the index has
exact spans.

The fix is to include the "include columns" when determining
whether the index has exact spans.

Change-Id: I6aca29358f40abcf7521b75bc83caedc90f14809
Reviewed-on: https://review.couchbase.org/c/query/+/217068
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>