MB-63656 MB-60680 Fix issues with index filter
If a vector index has include columns defined, then currently
we are not handling the include columns properly in generation
of index filters, specifically filters for the include columns.

The fix is to allow proper index filters to be generated for
include columns for both covering and non-covering index scans.

Change-Id: I30e799b9b80cabe8d857560bfe207ebdba21ba88
Reviewed-on: https://review.couchbase.org/c/query/+/217132
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>