MB-63773 Avoid inlining let expressions in cover expressions
This bug is a regression from the fix for MB-59642, where in
order to properly determine index covering and index key
projections in a query involving LET variables, we replace
LET variables in all query expressions before covering
determination. This replacement is supposed to be done on
a copy of the expressions, however, if there is a subquery
expression, since we don't actually copy the subquery
expression, we end up replacing LET variables with their
definitions in the subquery.

To fix this, we modify the previous fix, such that instead
of replacing LET variables, we already have the original
WHERE clause going through LET variable replacement early
on in the planning process (only the WHERE clause is
relevant in index selection). Thus we can simply add to
the expressions to be checked (which only includes the
original WHERE clause) the transformed WHERE clause as
well. Specifically:

 - when we replace LET variables in WHERE clause, also check
   whether the expression is modified, if so, set a new flag
   on the builder
 - when we gather expressions to be checked for covering, if
   the new builder flag is set, also includes the transformed
   WHERE clause
 - when setting identifier flag for UNNEST alias, also set
   the keyspace flag (this is similar to what happens in the
   formalizer where we set both flags
   in SetAllowedUnnestAlias() function). This change is
   required in case we have derived predicates for UNNEST in
   the transformed WHERE clause (e.g. <unnest_alias> IS NOT
   MISSING) the covering determination works properly
 - When we replace LET variables, instead of using the mapping
   from the LET definition itself, first make a copy. This is
   in case we have multiple levels of LET bindings, and since
   the inlining process just use whatever expression in the
   original LET definition when replacing an expression, if
   one LET binding references another, we could end up
   changing the LET binding expression itself

Change-Id: I8b91c9935191da90179e4dce293c86961757f39e
Reviewed-on: https://review.couchbase.org/c/query/+/217250
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>