MB-63702 Proper handling of optimizer hints in SubqueryTerm
Another issue I noticed with this defect is that if legacy-style
optimizer hints are used inside a SubqueryTerm, then the hint
status is not displayed properly in the explain output. It shows
"hints_status_unknown" instead of whether the hint is followed
or not followed.

For legacy-style optimizer hints (USE INDEX, USE NL, USE HASH)
the planner internally convert them into new derived optimizer
hints. This is done for a Subselect. For a SubqueryTerm though,
similar issue exists in that if the SubqueryTerm is planned
multiple times then we go through this process multiple times
as well, which causes issues.

The fix includes:

 - when deriving optimizer hints, instead of generating a new
   derived hint every time the Subselect is planned, first
   check whether a derived optimizer hint already exists, if
   so, use that instead of generating a new one
 - add capability for a few optimizer hint types to compare
   whether an existing hint is the same as one to be
   constructed. This is only needed for the types that has
   a legacy-style equivalent: HintIndex, HintFTSIndex, HintNL,
   HintHash.
 - add marking of keyspace hints for join enumeration if a
   SubqueryTerm is being planned, since the "normal" marking
   of optimizer hints are not done if join enumeration is in
   play.

Change-Id: I4de727c84e83e75860123248fad658a0f7370c1a
Reviewed-on: https://review.couchbase.org/c/query/+/217402
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Well-Formed: Restriction Checker